Fix templating/debug referring to the wrong bundle
The dump function documentation tells the user to install the VarDumper component, but this does not actually wire up the described function, leading to unexpected results when following the instructions (Twig's built-in dump function is used instead).

The actual wiring is done by DebugBundle, which requires VarDumper in turn. A simple change of requirement is all that's needed.